% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemap.R
\name{adjacency}
\alias{adjacency}
\title{Adjacency matrix from spacemap and spacemap models}
\usage{
adjacency(net, aszero = 1e-06)
}
\arguments{
\item{net}{Object output from \code{\link{spacemap}} and \code{\link{space.joint}}}

\item{aszero}{Positive numeric value (defaults to 1e-6) indicating at what point to consider extremely 
small parameter estimates of \eqn{\Gamma} and \eqn{\rho} as zero.}
}
\value{
A list containing 
\itemize{ 
 \item  \code{yy} A \eqn{p \times p} adjancency matrix 
 indicates an edge\eqn{y_q - y_l} when \code{yy[q,l] == 1} and 0 otherwise. 
 \item  \code{xy} A \eqn{p \times q} adjancency matrix 
 indicates an edge\eqn{x_p - y_q} when \code{xy[p,q] == 1} and 0 otherwise. 
}
}
\description{
Output from \code{\link{spacemap}} and \code{\link{space.joint}} is 
transformed into an adjacency matrix encoding a network.
}
