% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPerformance.R
\name{algoFDR}
\alias{algoFDR}
\title{False discovery rate}
\usage{
algoFDR(tp, fp)
}
\arguments{
\item{tp}{The number of true positive classifications.}

\item{fp}{The number of false positive classificationsm.}
}
\description{
Compute the FDR given true positives and false positives.
}
