% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPerformance.R
\name{algoMCC}
\alias{algoMCC}
\title{Matthew's Correlation Coefficient (MCC)}
\usage{
algoMCC(tp, fn, fp, tn)
}
\arguments{
\item{tp}{The number of true positive classifications.}

\item{fn}{The number of false negative classifications.}

\item{fp}{The number of false positive classifications.}

\item{tn}{The number of true negative classifications.}
}
\description{
Compute the MCC given the confusion matrix.
}
