% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPerformance.R
\name{algoPower}
\alias{algoPower}
\title{Statistical Power of an Algorithm}
\usage{
algoPower(tp, fn)
}
\arguments{
\item{tp}{The number of true positive classifications.}

\item{fn}{The number of false negative classifications.}
}
