% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bicCondLike}
\alias{bicCondLike}
\title{BIC for space under conditional chain graph likelihood setting.}
\usage{
bicCondLike(fit, tol, Y, X, model = c("spacemap", "space", "scggm"),
  Yindex = NULL, Xindex = NULL)
}
\arguments{
\item{fit}{The fitted model object returned from spacemap::[spacemap, space.joint].}

\item{tol}{A numeric lower bound on non-zero parameter values.}

\item{Y}{A numeric matrix of N X Q responses.}

\item{X}{A numeric matrix of N X P predictors.}

\item{NN}{An integer, the sample size.}
}
\value{
the Bayesian Information Criterion (BIC) under conditional chain graph likelihood formulation.
}
\description{
Calculate the BIC under the conditional chain graph likelihood.
}
\references{
Zhang L, Kim S (2014) Learning Gene Networks under SNP Perturbations Using eQTL Datasets. 
PLoS Comput Biol 10(2): e1003420. doi: 10.1371/journal.pcbi.1003420
}
