% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bicLike}
\alias{bicLike}
\title{BIC for space under likelihood setting.}
\usage{
bicLike(fit, tol, dat)
}
\arguments{
\item{fit}{The returned object from spacemap::space.}

\item{tol}{A numeric lower bound on non-zero parameter values.}

\item{dat}{the data matrix.}
}
\value{
the Bayesian Information Criterion (BIC).
}
\description{
Calculate the BIC.
}
