% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossValidation.R
\name{cvPerf}
\alias{cvPerf}
\title{CV.Vote Model Performance}
\usage{
cvPerf(cvOut, truth, method, givenX = FALSE, Xindex = NULL, Yindex = NULL,
  aszero = 1e-06)
}
\arguments{
\item{cvOut}{Output from \code{\link{cvVote}}.}

\item{truth}{A list of numeric matrices named 'xy' and 'yy' 
encoding the true model parameters (or an adjacency matrix).}

\item{method}{Character vector indicates \code{cvOut} came from
spaceMap or SPACE model. See parameter \code{method} in \code{\link{cvVote}}.}

\item{givenX}{Logical (defaults to FALSE). Set to TRUE for \code{method == "space"}
when \code{cvOut} had non-null \code{X} input to \code{\link{cvVote}}.}

\item{Xindex}{Integer vector of indices of X variables in partial correlation matrix. Defaults to 
NULL, but must be non-null when \code{method == "space" & givenX = TRUE}.}

\item{Yindex}{Integer vector of indices of Y variables in partial correlation  matrix. Defaults to 
NULL, but must be non-null when \code{method == "space" & givenX = TRUE}.}

\item{aszero}{A numeric value specifying the point at which a parameter estimate should 
be effectively considered of zero value.}
}
\description{
Convenience function for evaluating edge detection performance of 
CV.Vote model fits.
}
\seealso{
\code{\link{reportJointPerf}}, \code{\link{reportPerf}}
}
