% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initFit.R
\name{initFit}
\alias{initFit}
\title{Identify scale of tuning penalties}
\usage{
initFit(Y, X = NULL, tuneGrid, method = c("spacemap", "space"),
  iscale = TRUE, aszero = 1e-06, ...)
}
\arguments{
\item{Y}{Numeric matrix \eqn{(N \times Q)} containing N iid samples of the response vector \eqn{\textbf{y}}.}

\item{X}{Numeric matrix \eqn{(N \times P)} containing N iid samples of the predictor vector \eqn{\textbf{x}}.}

\item{tuneGrid}{Named with columns \code{lam1, lam2, lam3} 
when \code{method = "spacemap"}. Each row in the data.frame corresponds to a
 tuning parameter set that is input into \code{\link{spacemap}}. 
When \code{method = "space"}, supply a data.frame with only one column 
being \code{lam1}.}

\item{method}{Character vector indicates network inference with function 
\code{\link{spacemap}} when \code{method = "spacemap"} or function 
\code{\link{space.joint}} when \code{method = "space"}. If \code{X} is 
non-null and \code{method = "space"}, then \code{space.joint} will 
infer (x--x, x--y, y--y) edges but only report (x--y, y--y) edges.}

\item{iscale}{Logical indicating to standardize the whole input data. Defaults to TRUE. 
See \code{\link{base::scale(x, center = TRUE, scale = TRUE)}} for details of standardization.}

\item{aszero}{Positive numeric value (defaults to 1e-6) indicating at what point to consider extremely 
small parameter estimates of \eqn{\Gamma} and \eqn{\rho} as zero.}

\item{...}{Additional arguments for \code{\link{spacemap}} or \code{\link{space.joint}}
to change their default settings (e.g. setting \code{tol = 1e-4}).}
}
\value{
If \code{method=="spacemap"} or (\code{method=="space"} and \code{X!=NULL}),
 return a  data.frame where the first column \code{nyy} reports the 
 number of \eqn{y-y} edges  and the second column \code{nxy} reports 
 the number of \eqn{x-y} edges. Rows of the data.frame correspond to 
 the input parameter \code{tuneGrid}. 
 
 If \code{method=="space"} and \code{X==NULL}, return a vector 
 of the number of \eqn{y-y} edges.
}
\description{
When initially choosing tuning penalties, it can be 
challenging to find the appropriate scale. This function fits either 
\code{\link{spacemap}} or \code{\link{space.joint}} once for a specified 
tuning grid on the whole data. It reports the corresponding number of
\eqn{y-y} edges and \eqn{x-y} edges for each tuning penalty set. 
Having a prior understanding of how sparse the network ought to 
be can help narrow the scale of the tuning grid 
based on the output of this function.
}
\seealso{
\code{\link{cvVote}}
}
