% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nonZeroUpper}
\alias{nonZeroUpper}
\title{Count the number of y-y edges}
\usage{
nonZeroUpper(ParCor, aszero)
}
\arguments{
\item{ParCor}{The estimated partial correlation matrix, \eqn{\hat\rho}, 
output from \code{\link{spacemap}}. Assumed to be symmetric.}

\item{aszero}{Positive numeric value indicating at what point to consider extremely 
small parameter estimates of \eqn{\rho} as zero.}
}
\value{
Number of y-y edges.
}
\description{
Count the number of y-y edges after fitting spaceMap model.
}
