% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nonZeroWhole}
\alias{nonZeroWhole}
\title{Count the number of x-y edges}
\usage{
nonZeroWhole(Gamma, aszero)
}
\arguments{
\item{Gamma}{The estimated regression coefficient matrix, \eqn{\hat\Gamma}, 
output from \code{\link{spacemap}}.}

\item{aszero}{Positive numeric value indicating at what point to consider extremely 
small parameter estimates of \eqn{\Gamma} as zero.}
}
\value{
Number of x-y edges.
}
\description{
Count the number of x-y edges after fitting spaceMap model.
}
