% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networkAnalysis.R
\name{reportHubs}
\alias{reportHubs}
\title{Report top hub nodes with attributes.}
\usage{
reportHubs(ig, top = 10, level = c("x", "y"), unit = "Mb",
  unit_base = 1e+06, sig = 2)
}
\arguments{
\item{ig}{An igraph network object returned from either 
\code{\link{rankHub}} or \code{\link{cisTrans}}.}

\item{top}{An integer specifying the number of top-degree hubs to report.}

\item{level}{Character specifying nodes of level x or y, but not both. 
Defaults to x.}

\item{unit}{If gene coordinates for x-hubs exist, then this will 
label the chosen genome location scale in the table. Defaults to mega-base (Mb).}

\item{unit_base}{Numeric divisor of the genome coordinates and should 
correspond to \code{unit} label. Defaults to 1e6 for mega-base.}

\item{sig}{Integer for controlling precision of genome coordinates.}
}
\value{
A table of top ranked node hubs with attributes deemed 
relevant to report. See vignette on network analysis for examples.
}
\description{
Organize x and y hub analysis into a publish-ready format.
}
\seealso{
\code{\link{rankHub}}, \code{\link{cisTrans}}, 
\code{\link{adj2igraph}}, \code{\link{xHubEnrich}}, \code{\link{modEnrich}}
}
