% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPerformance.R
\name{reportJointPerf}
\alias{reportJointPerf}
\title{Report joint performance of (x-y,y-y) edge detection}
\usage{
reportJointPerf(fit, truth, aszero = 1e-06, verbose = TRUE)
}
\arguments{
\item{fit}{A list of numeric matrices named 'xy' and 'yy' 
encoding the fitted model parameter estimates (or the true adjacency matrices).}

\item{truth}{A list of numeric matrices named 'xy' and 'yy' 
encoding the true model parameters (or an adjacency matrix).}

\item{aszero}{A numeric value specifying the point at which a parameter estimate should 
be effectively considered of zero value.}

\item{verbose}{A logical value indicating whether to report the performance to the console.}
}
\value{
A numeric, named vector 
\itemize{
\item{power}{ Joint power on (x-y,y-y)}
\item{fdr}{ Joint false discovery rate (FDR) on (x-y,y-y)}
\item{powerXY}{ Power of x-y edges}
\item{fdrXY}{ FDR of x-y edges}
\item{powerYY}{ Power of y-y edges}
\item{fdrYY}{ FDR of y-y edges}
\item{tp}{ True positives of (x-y,y-y)}
\item{fn}{ False negatives of (x-y,y-y)}
\item{fp}{ False positives of (x-y,y-y)}
\item{tpXY}{ True positives of x-y}
\item{fnXY}{ False negatives of x-y}
\item{fpXY}{ False positives of x-y}
\item{tpYY}{ True positives of y-y}
\item{fnYY}{ False negatives of y-y}
\item{fpYY}{ False positives of y-y}
}
}
\description{
Assesses the network learning performance of spaceMap (or space)
against a known truth by reporting power and false discovery rate.
}
\seealso{
reportPerf
}
