% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodEdge.R
\name{reprodEdge}
\alias{reprodEdge}
\alias{bootVote}
\title{Robust networks through bootstrap}
\usage{
reprodEdge(Y.m, X.m, tune, boot = TRUE, B = 1000L, seed = 55139, p0 = 1,
  iter = 3, tol = 1e-04, cd_iter = 1e+09, refitRidge = 0.1)

bootVote(fits, P, Q, vote_thresh = 0.5, btol = 0)
}
\arguments{
\item{Y.m}{A numeric matrix (N x Q) representing the response of Q variables on 
N independent observations.}

\item{X.m}{A numeric matrix (N X P) representing P predictor variables on N 
independent observations.}

\item{iter}{An integer specifying the iterations in JSRM for estimating 
\eqn{\sigma^{ii}} and partial correlations. When \code{sig=NULL} and/or 
(\code{weight=NULL} or \code{weight=2}), then \code{iter} \eqn{\ge 2}.}

\item{fits}{List of fitted spacemap models returned from \code{\link{reprodEdge}}.}

\item{P}{the number of predictor variables}

\item{Q}{the number of response variables}

\item{vote_thresh}{the threshold for the minimum proportion of bootstrap 
replicate model fits with a particular edge such that the edge is included in the Boot.Vote model .}

\item{btol}{the threshold for the smallest possible parameter estimate in the model 
to be counted as an edge in a model fit. Do not change from zero if used in tandem 
with \code{\link{reprodEdge}}.}
}
\description{
The Boot.Vote procedure fits spacemap over B bootstrap replicates
 for each tuning parameter set of (\code{slasso,rlasso,rgroup}).

Aggregate boostrap replicates of spacemap into a final Boot.Vote model.
}
\examples{
\dontrun{
data(sim1)
tune <- data.frame(slasso = 70, rlasso = 28.8, rgroup = 14)
out <- reprodEdge(Y.m = dat$Y, X.m = dat$X, tune = tune, B = 50, iter = 3,
                 cd_iter = 1e7, tol = 1e-6, refitRidge = 0)
bv <- bootVote(fits = out[[1]], P = P, Q = Q)
}
}
