% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simRemMap.R
\name{simRemMap}
\alias{simRemMap}
\title{Simulate remMap scenario.}
\usage{
simRemMap(n = 100, p = 300, q = 300, seed = 1)
}
\description{
Generate data suitable for remMap.
}
\details{
Wrapper around example from remMap documentation.
}
