% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space.joint}
\alias{space.joint}
\title{Estimate partial correlations using the Joint Sparse Regression Model}
\usage{
See space::space.joint
}
\description{
Estimate partial correlations using the Joint Sparse Regression Model
}
