% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/space.R
\name{space}
\alias{space}
\title{Fit the SPACE model.}
\usage{
space(Y, lam1, sridge = 0, sig = NULL, weight = NULL, iter = 3,
  tol = 1e-06, cdmax = 10000000L, rho = NULL, iscale = TRUE)
}
\arguments{
\item{Y}{Numeric matrix \eqn{(N \times Q)} containing N iid samples of the response vector \eqn{\textbf{y}}.}

\item{lam1}{Non-negative numeric value, the space-lasso penalty corresponding to \eqn{\lambda_1}, 
which subjects the partial correlation vector, \eqn{\bf \rho_{yy}}, to  the \eqn{l_1} norm. 
It induces overall sparsity of \eqn{\{ y_q - y_l : q \neq l \}} edges. .}

\item{sridge}{The \eqn{l_2} penalty parameter defaults to 0. When it is positive, 
it imposes an elastic net  penalty in tandem with \code{lam1} on \eqn{\rho}.}

\item{sig}{Positive numeric vector (\eqn{p \times 1}) representing the estimate of \eqn{\sigma^{ii}}, the diagonal of 
the inverse covariance matrix. It defaults to NULL and
and will be estimated \code{iter} times during the model fitting with initial 
values set to the ones vector.}

\item{weight}{The weights applied to the Q regressions in the SPACE model. 

 If \code{weight==NULL}, each regression has equal weight.
 
 If \code{weight==1}, each regression is weighted according to 
 \eqn{\sigma^{ii}} and enforces \code{iter>=2}.
 
 If \code{weight==2} each regression is weighted according to
  the out-degree of \eqn{q}th response variable and enforces \code{iter>=2}.
 
 Otherwise, \code{weight=}user-specified vector of length p. 

The scale of weight does not matter. In this function, weight will be rescaled to have mean=1}

\item{iter}{Positive integer specifying the number of iterations for estimating 
\eqn{\sigma^{ii}} and \eqn{\rho}. Defaults to 3.}

\item{tol}{Positive numeric value specifying the convergence tolerance of the coordinate descent algorithm; 
in other words, it is criterion that stops parameter estimation when no parameter changes value exceeding \code{tol} 
between iterations. \code{tol} defaults to 1e-6, but may be lowered (e.g. 1e-4) to speed
up network learning.}

\item{cdmax}{Positive integer specifiying the maximum number of parameter updates allowed before reporting the algorithm
as having failed to converge. Default may need to be increased for inferring very large-scale networks (i.e. \eqn{p,q > 1000}).}

\item{rho}{Numeric matrix (\eqn{p \times p}) representing the estimate of \eqn{\hat\rho}, the partial correlation matrix. 
It defaults to NULL and and will be estimated \code{iter} times during the model fitting.}

\item{iscale}{Logical indicating to standardize the whole input data. Defaults to TRUE. 
See \code{base::scale(x, center = TRUE, scale = TRUE)} for details of standardization.}
}
\value{
A list containing 
\enumerate{
  \item \code{ParCor} The estimated partial correlation matrix (\eqn{P \times P}), 
  where off-diagonals  \eqn{ |\hat \rho^{p,q}_{yy}| > 1e-6}  encode the edges \eqn{\{ y_q - y_l : q \neq l \} }
  and the diagonals are 1's. 
  \item \code{sig.fit} The estimated diagonal \eqn{\hat \sigma^{ii}}.
  \item \code{rss} The residual sums of squares from the model fit. 
  \item convergence logical: true for successful convergence, otherwise failed to converge. Failure can be 
  mitigated by increasing \code{tol} and/or \code{cdmax}. 
  \item \code{deltaMax} The maximum change in parameter values between the penultimate and ultimate iteration. 
  If \code{spacemap} does not converge, \code{deltaMax} provides some measure of how far away it was from converging
  when compared to \code{tol}. 
}
}
\description{
Estimate partial correlations according to SPACE model.
}
\examples{
data(sim1)
net <- spacemap::space(Y = sim1$Y, lam1 = 70)
#adjacency matrix of y-y edges. 
adjnet <- adjacency(net)
}
\seealso{
\code{\link{spacemap}}, \code{\link{cvVote}}, \code{\link{bootEnsemble}}, \code{\link{bootVote}}
}
