% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacemap.R
\name{spacemap}
\alias{spacemap}
\title{Fit spacemap model}
\usage{
spacemap(Y.m, X.m, slasso, sridge = 0, rlasso, rgroup, sig = NULL,
  weight = NULL, remWeight = NULL, iter = 3, tol = 1e-06,
  cd_iter = 10000000L, verbose = FALSE)
}
\arguments{
\item{Y.m}{A numeric matrix (N x Q) representing the response of Q variables on 
N independent observations.}

\item{X.m}{A numeric matrix (N X P) representing P predictor variables on N 
independent observations.}

\item{slasso}{A numeric value, the Space-lasso penalty parameter (l_1 norm).}

\item{sridge}{A numeric value (optional). If not specified, lasso regression 
is used in the Joint Sparse Regression Model (JSRM). Otherwise, elastic net 
regression is used in JSRM and sridge serves as the squared l_2 norm penalty parameter.}

\item{rlasso}{The MAP-lasso penalty.}

\item{rgroup}{The MAP-group penalty.}

\item{sig}{A numeric vector with length being the same as the number of 
columns of Y.m. It is the vector of \eqn{\sigma^{ii}} (the diagonal of 
the inverse covariance matrix). If not specified, σ^{ii} will be 
estimated during the model fitting with initial values \code{rep(1,p)} and  
the iterations of the model fitting (\code{iter}) \eqn{\ge 2}. 
Note, the scale of \code{sig} does not matter.}

\item{weight}{A numeric value or vector (Q x 1) specifying the weights or the type of
weights used for each regression in JSRM. The default value is NULL, 
which means all regressions will be weighted equally in the joint model. 
If \code{weight=1}, residue variances will be used for weights. 
If \code{weight=2}, the estimated degree of each variable will be used for weights. 
Otherwise, it should be a positive numeric vector, whose length is equal to 
the number of columns of Y.m.}

\item{remWeight}{optional weighting scheme on [fill in details].}

\item{iter}{An integer specifying the iterations in JSRM for estimating 
\eqn{\sigma^{ii}} and partial correlations. When \code{sig=NULL} and/or 
(\code{weight=NULL} or \code{weight=2}), then \code{iter} \eqn{\ge 2}.}
}
\value{
A list containing 
\enumerate{
  \item \code{ParCor} The estimated partial correlation matrix. 
  \item \code{sig.fit} The estimated diagonal \eqn{\sigma^{ii}}.
  \item \code{Gamma} The estimated coefficients of Y on X.
  \item \code{rss} The residual sums of squares from the model fit. 
  \item \code{deltaMax} The maximum change in parameter values between the penultimate and ultimate iteration  
}
}
\description{
Estimate partial correlations between response variables (Y)
and perform model selection of predictors (X) based on the spacemap model 
fitted to multivariate data (Y,X).
}
\examples{
data(sim1)
net <- spacemap(Y.m = sim1$Y, X.m = sim1$X, slasso = 70, rlasso = 28.8, rgroup = 12.38)
}
