% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneVis.R
\name{tuneVis}
\alias{tuneVis}
\title{Visualize cross validation metrics}
\usage{
tuneVis(cvOut, testSetLen, tuneParam1, tuneParam1Name = c("lam1", "lam2",
  "lam3"), tuneParam2 = NULL)
}
\arguments{
\item{cvOut}{Output from \code{\link{cvVote}}.}

\item{testSetLen}{Integer vector specifying length of each test set associated with 
\code{cvOut}.}

\item{tuneParam1}{Numeric vector specifying the tuning penalty placed on the 
x-axis.}

\item{tuneParam1Name}{Character vector labelling the X-axis with the specified tuning parameter.}

\item{tuneParam2}{Numeric vector specifying the tuning penalty placed on the 
x-axis. Defaults to NULL. If non-null, the output is a heatmap where 
CV scores are colored intensities on a 2-D grid of \code{tuneParam1,tuneParam2}.}
}
\value{
If \code{tuneParam2 == NULL}
Return a list of 4 \code{ggplot2} objects. 
\enumerate{
\item Scatter plot of log (CV score) vs. \code{tuneParam1}. 
\item Scatter plot of "Avg. Total # of Edges" vs. \code{tuneParam1}. 
\item Scatter plot of "Avg. # of Edges (Y -- Y)" vs. \code{tuneParam1}.
\item Scatter plot of "Avg. # of Edges (x -> Y)" vs. \code{tuneParam1}.
}

If \code{tuneParam2 != NULL}
Returns a heatmap (see \code{ggplot2::geom_tile}) where 
CV scores are colored intensities on a 2-D grid of \code{tuneParam1,tuneParam2}.
}
\description{
Diagnose the quality of the tuning grid used in \code{\link{cvVote}}
through plotting  CV scores and network size against the
tuning penalties.
}
\seealso{
\code{\link{cvVote}}
}
