% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPerformance.R
\name{algoMCC}
\alias{algoMCC}
\title{Mathew's Correlation Coefficient of Classifier}
\usage{
algoMCC(tp, fn, fp, tn)
}
\arguments{
\item{tp}{The number of true positive classifications.}

\item{fn}{The number of false negative classifications.}

\item{fp}{The number of false positive classifications.}

\item{tn}{The number of true negative classifications.}
}
