% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodEdge.R
\name{bootEnsemble}
\alias{bootEnsemble}
\title{Ensemble networks through boostrapping.}
\usage{
bootEnsemble(Y, X = NULL, tune, method = c("spacemap", "space"),
  boot = TRUE, B = 100L, resPath = tempdir(), iscale = TRUE,
  aszero = 1e-06, seed = sample.int(n = 1e+06, size = 1), p0 = 1, ...)
}
\arguments{
\item{Y}{Numeric matrix \eqn{(N \times Q)} containing N iid samples of the response vector \eqn{\textbf{y}}.}

\item{X}{Numeric matrix \eqn{(N \times P)} containing N iid samples of the predictor vector \eqn{\textbf{x}}.}

\item{tune}{List with names being the tuning penalties
\code{lam1, lam2, lam3} and each element being a single numeric value that is input to \code{spacemap}. One might  
consider the output from \code{\link{cvVote}} (i.e. list \code{minTune}) as input for this parameter.}

\item{method}{Character value specifying to use \code{\link{spacemap}} or 
\code{\link{space}} in fitting ensemble network.}

\item{boot}{Logical. Default is \code{boot=TRUE} and implies sampling with replacement (i.e. boostrap). 
\code{boot=FALSE} will resample without replacement.}

\item{B}{Positive integer denoting the number of model fits making up the ensemble. Default is 1000, but can 
be lowered to save time.}

\item{resPath}{Character vector specifying the directory where each 
each model fit to bootstrap replicate data is written to file through serialization by \code{saveRDS}. 
Defaults to temporary directory that will be deleted at end of the R session. 
It is recommended to specify a directory where results can be stored permanently.}

\item{iscale}{Logical indicating to standardize the whole input data. Defaults to TRUE. 
See \code{base::scale(x, center = TRUE, scale = TRUE)} for details of standardization.}

\item{aszero}{Positive numeric value (defaults to 1e-6) indicating at what point to consider extremely 
small parameter estimates of \eqn{\Gamma} and \eqn{\rho} as zero.}

\item{seed}{Positive integer allowing user to set the random seed for reproducing the network ensemble.}

\item{p0}{Positive numeric not exceeding the default of 1, which represents the proportion of the original 
samples that will be down-sampled for each model fitting iteration of the ensemble.}

\item{...}{Additional arguments for \code{\link{spacemap}} or  \code{\link{space}}.}
}
\value{
A list, call it \code{ens}, up to length \code{B} of convergent model fits from 
either \code{\link{spacemap}} or \code{\link{space}}. \code{ens} should typically
not be modified by the user, but passed on to the \code{\link{bootVote}} function.
Non-convergent model fits to bootstrap replciates are not included in list `ens`, 
but can be inspected individually at location \code{resPath} and are identified  
by \code{attr(good, which = "not_conv_bootids")}.
}
\description{
Fits spacemap model to B bootstrapped data replicates to create an ensemble network.
}
\examples{

#Load simulation
library(spacemap)
data(sim1)

#Boostrap Ensemble (B = 10) for illustration only
tune <- data.frame(lam1 = 70, lam2 = 28, lam3 = 17.5)
ens <- bootEnsemble(Y = sim1$Y, X = sim1$X, tune = tune, 
                    method = "spacemap", B = 10)
}
\seealso{
\code{\link{bootVote}}, \code{\link{cvVote}}
}
