% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reprodEdge.R
\name{bootEnsemble}
\alias{bootEnsemble}
\title{Ensemble networks through boostrapping.}
\usage{
bootEnsemble(Y, X = NULL, tune, method = c("spacemap", "space"),
  boot = TRUE, B = 1000L, iscale = TRUE, aszero = 1e-06,
  seed = sample.int(n = 1e+06, size = 1), p0 = 1, ...)
}
\arguments{
\item{Y}{Numeric matrix \eqn{(N \times Q)} containing N iid samples of the response vector \eqn{\textbf{y}}.}

\item{X}{Numeric matrix \eqn{(N \times P)} containing N iid samples of the predictor vector \eqn{\textbf{x}}.}

\item{tune}{List with names being the tuning penalties
\code{lam1, lam2, lam3} and each element being a single numeric value that is input to \code{spacemap}. One might  
consider the output from \code{\link{cvVote}} (i.e. list \code{minTune}) as input for this parameter.}

\item{boot}{Logical. Default is \code{boot=TRUE} and implies sampling with replacement (i.e. boostrap). 
\code{boot=FALSE} will resample without replacement.}

\item{B}{Positive integer denoting the number of model fits making up the ensemble. Default is 1000, but can 
be lowered to save time.}

\item{iscale}{Logical indicating to standardize the whole input data. Defaults to TRUE. 
See \code{\link{base::scale(x, center = TRUE, scale = TRUE)}} for details of standardization.}

\item{aszero}{Positive numeric value (defaults to 1e-6) indicating at what point to consider extremely 
small parameter estimates of \eqn{\Gamma} and \eqn{\rho} as zero.}

\item{seed}{Positive integer allowing user to set the random seed for reproducing the network ensemble.}

\item{p0}{Positive numeric not exceeding the default of 1, which represents the proportion of the original 
samples that will be down-sampled for each model fitting iteration of the ensemble.}
}
\value{
A list up to length \code{B} of convergent and sparse model fits from 
either \code{\link{spacemap}} or \code{\link{space}}. The list object should typically
not be modified by the user, but passed on to the  \code{\link{bootVote}} function.
Model fits that are not convergent or produce networks that are not sparse are not reported.
}
\description{
Fits spacemap model to B bootstrapped data replicates to create an ensemble network.
}
\examples{

#Load simulation
data(sim1)

#Boostrap Ensemble (B = 10) for illustration only
tune <- data.frame(lam1 = 70, lam2 = 28, lam3 = 17.5)
ens <- bootEnsemble(Y = sim1$Y, X = sim1$X, tune = tune, 
                    method = "spacemap", B = 10)
}
\seealso{
\code{\link{bootVote}}, \code{\link{cvVote}}
}
